+++ crates/uiautomation/src/core.rs
@@ -1663,6 +1663,7 @@ mod tests {
     use crate::UIElement;
     use crate::controls::ControlType;
     use crate::filters::MatcherFilter;
+    use crate::types::Handle;
     use crate::types::TreeScope;
 
     fn print_element(element: &UIElement) {
@@ -1832,5 +1833,11 @@ mod tests {
         
         let uiautomation = UIAutomation::new_direct();
         assert!(uiautomation.is_ok());
+    }
+
+    #[test]
+    fn test_search_from_handle() {
+        let auto = UIAutomation::new().unwrap();
+        let _ = auto.element_from_handle(Handle::from(0x2006C6));
     }
 }