+++ CHANGELOG.md
@@ -102,3 +102,11 @@
 ## v0.6.4
 
 + `Keyboard.begin_hold_keys()` & `Keyboard.end_hold_keys()` respond to `Keyboard.interval()` setting.
+
+## v0.7.0
+
++ refactor `Process`, add more controls.
+
+## v0.7.2
+
++ support converting `isize` to `Handle`. [issue-44](https://github.com/leexgone/uiautomation-rs/issues/44)